add node parser pkg, expand READMEs & examples
