+++ llama-index-integrations/node_parser/llama-index-node-parser-docling/llama_index/node_parser/docling/__init__.py
@@ -1,0 +1,4 @@
+from llama_index.node_parser.docling.base import DoclingNodeParser
+
+
+__all__ = ["DoclingNodeParser"]
