+++ llama-index-integrations/node_parser/llama-index-node-parser-docling/pyproject.toml
@@ -11,7 +11,7 @@ skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"
 
 [tool.llamahub]
 contains_example = true
-import_path = "llama_index.readers.docling"
+import_path = "llama_index.node_parser.docling"
 
 [tool.llamahub.class_authors]
 DoclingPDFReader = "vagenas"
@@ -25,9 +25,9 @@ python_version = "3.10"
 
 [tool.poetry]
 authors = ["Panos Vagenas <pva@zurich.ibm.com>"]
-description = "llama-index readers docling integration"
+description = "llama-index node_parser docling integration"
 license = "MIT"
-name = "llama-index-readers-docling"
+name = "llama-index-node-parser-docling"
 packages = [{include = "llama_index/"}]
 readme = "README.md"
 version = "0.1.0"
@@ -36,7 +36,6 @@ version = "0.1.0"
 python = "^3.10"
 llama-index-core = "^0.11.0"
 docling-core = {branch = "align-chunk-path-format", git = "https://github.com/DS4SD/docling-core.git"}  # TODO temp
-docling = "^1.16.1"
 
 [tool.poetry.group.dev.dependencies]
 black = {extras = ["jupyter"], version = "<=23.9.1,>=23.7.0"}
@@ -60,5 +59,7 @@ types-setuptools = "67.1.0.0"
 [tool.poetry.group.examples.dependencies]
 llama-index-embeddings-huggingface = "^0.3.1"
 llama-index-llms-huggingface-api = "^0.2.0"
+llama-index-readers-docling = {develop = true, path = "../../readers/llama-index-readers-docling/"}  # TODO temp
 llama-index-vector-stores-milvus = "^0.2.5"
 python-dotenv = "^1.0.1"
+rich = "^13.9.1"
